# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Soccer RESTful API Documentation",
        "description": "Dokumentasi API untuk Soccer RESTful API (Laravel) menggunakan L5 Swagger.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/pertandingan": {
            "get": {
                "tags": [
                    "Pertandingan"
                ],
                "summary": "Menampilkan daftar pertandingan",
                "operationId": "35af0d44f3db31a1a1208c0179de771f",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil daftar pertandingan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Pertandingan"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Access Denied"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Terjadi kesalahan pada server."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tidak ada data pertandingan ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Tidak ada data pertandingan ditemukan"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Pertandingan"
                ],
                "summary": "Menambah pertandingan baru",
                "operationId": "93243ed48e097b0e55d85f992714b0e4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tanggal",
                                    "waktu",
                                    "lokasi",
                                    "tim_home_id",
                                    "tim_away_id",
                                    "liga_id"
                                ],
                                "properties": {
                                    "tanggal": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-08-01"
                                    },
                                    "waktu": {
                                        "type": "string",
                                        "example": "15:30"
                                    },
                                    "lokasi": {
                                        "type": "string",
                                        "example": "Stadion Utama"
                                    },
                                    "tim_home_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "tim_away_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "skor_home": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "skor_away": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "liga_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pertandingan berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Pertandingan"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Access Denied"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Terjadi kesalahan pada server."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Pertandingan": {
                "title": "Pertandingan",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "tanggal": {
                        "type": "string",
                        "format": "date",
                        "example": "2023-08-01"
                    },
                    "waktu": {
                        "type": "string",
                        "example": "15:30"
                    },
                    "lokasi": {
                        "type": "string",
                        "example": "Stadion Utama"
                    },
                    "tim_home_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "tim_away_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "skor_home": {
                        "type": "integer",
                        "example": 2
                    },
                    "skor_away": {
                        "type": "integer",
                        "example": 1
                    },
                    "liga_id": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Masukkan token dengan format: Bearer {token}",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}